refactoring the index so it does the adjustment for the frontend for the animals and updating lunar visibility to display visible or not visible
