+++ src/react/ui/pages/lunar/lunarVisibilityPage/lunarVisibilityPageViewModel.ts
@@ -5,27 +5,21 @@ export const useLunarVisibilityPageViewModel = () => {
   const [lunarCount, setLunarCount] = useState([]);
   const [animalList, setAnimalList] = useState([]);
 
-  const adjustLunarVisibilityCount = (list: {isMoonVisible: boolean, animalName: string, callCount: number}[]):void => {
-    const newList = [];
-    var currentObj = {};
-    list.forEach((animalCall: {isMoonVisible: boolean, animalName: string, callCount: number}) => {
-      const isMoonVisible = animalCall.isMoonVisible ? "visible" : "not visible";
-      if (currentObj?.isMoonVisible !==  isMoonVisible) {
-        if(Object.keys(currentObj).length > 0) {newList.push(currentObj)}
-        currentObj = { isMoonVisible: isMoonVisible, total: 0};
-      }
-      const name = animalCall.animalName.replaceAll(" ", "");
-      currentObj[name] = animalCall.callCount;
-      currentObj.total += animalCall.callCount;
-      
-    });
-    newList.push(currentObj);
-    setLunarCount(newList)
+  const updateVisibleStatus = (data: []): any[] => {
+    const newMap = []
+    data.forEach((item) => {
+      let newItem = structuredClone(item);
+      newItem.isMoonVisible = item.isMoonVisible ? "visible" : "not visible";
+      newMap.push(newItem);
+    })
+    return newMap;
+
   }
 
   const fetchLunarVisibilityCount = (): void => {
     const lunarList = window?.electron?.sendSync('get-callCount', {page: CountPage.LunarVis});
-    adjustLunarVisibilityCount(lunarList);
+    const updatedList = updateVisibleStatus(lunarList);
+    setLunarCount(updatedList)
   }
 
   const fetchAnimalSpecies = ():void => {
